requirements.txt
dspy-ai
google-generativeai
